+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.1.4
+
+- fix: only trigger `onChanged` in `StringField` when the value actually changes
+
 ## 0.1.3
 
 - feat: add `DoubleField` `DateField`, `TimeField`, `DateTimeField`